% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vigenerecipher.R
\name{vigdecrypt}
\alias{vigdecrypt}
\title{Function to decrypt a message using the Vigenere cipher.}
\usage{
vigdecrypt(message, key)
}
\arguments{
\item{message}{A string containing the message to be decrypted in upper case.}

\item{key}{A string containing the keyword in upper case.}
}
\value{
The decrypted value of the message using the Vigenere cipher with the
specified keyword.
}
\description{
Function to decrypt a message using the Vigenere cipher.
}
\details{
Given a \code{message} and a \code{key}, this function decrypts the
message. NOTE: if \code{message} or \code{key} contain any non-alphabetic
characters, an error is raised.
}
\examples{

vigdecrypt('UMIVZFMLVFTIQGX', 'MATRIX')
# returns 'imperialcollege'

}
