% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vigenerecipher.R
\name{vigencrypt}
\alias{vigencrypt}
\title{Function to encrypt a message using the Vigenere cipher.}
\usage{
vigencrypt(message, key)
}
\arguments{
\item{message}{A string containing the message to be encrypted in lower
case.}

\item{key}{A string containing the keyword in upper case.}
}
\value{
The encrypted value of the message using the Vigenere cipher with the
specified keyword.
}
\description{
Function to encrypt a message using the Vigenere cipher.
}
\details{
Given a \code{message} and a \code{key}, this function encrypts the
message. NOTE: if \code{message} or \code{key} contain any non-alphabetic
characters, an error is raised.
}
\examples{

vigencrypt('imperialcollege', 'MATRIX')
# returns 'UMIVZFMLVFTIQGX'

}
